Fix double-counting of commits
When committer and author are distinguished (as is the
case for proximity, committer2author, and the file based
analysis modes), both persons are associated with one given
commit. However, both roles can be assumed by one person,
which leads to a double-counting of commit information in
the current implementation. Fix this.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@oth-regensburg.de>
Reported-by: 叶德铭 <ydm14@mails.tsinghua.edu.cn>
Tested-by: 李明杰 <li-mj14@mails.tsinghua.edu.cn>